+++ lib/ui/calender_view/calender_monthly_view.dart
@@ -59,6 +59,7 @@ class _CalenderMonthlyViewState extends State<CalenderMonthlyView> {
   }
 
   init() async {
+    _instance.calculateLastPeriodDate();
     pastAllPeriodsDays = _instance.pastAllPeriodDays;
     futurePeriodDays = await initFuturePeriodDay();
     futureOvulationDays = await initFutureOvulationDay();